move translations file
