+++ frontend/chat-plugin/lib/src/translations.ts
